#152: fix logger in tests
