+++ apps/backend/test/vitest.setup.ts
@@ -6,6 +6,7 @@ import { fileURLToPath } from 'url';
 import { vi } from 'vitest';
 import { migrateToLatest } from '../src/db/migrateToLatest.js';
 import type { DB } from '../src/db/types/db.js';
+import { kyselyLogger } from '../src/logger.js';
 
 const FILENAME = fileURLToPath(import.meta.url);
 const DIRNAME = path.dirname(FILENAME);
@@ -18,6 +19,7 @@ const migrationClient = new Kysely<DB>({
       connectionString: container.getConnectionUri(),
     }),
   }),
+  log: kyselyLogger,
 });
 
 const migrationPath = path.join(DIRNAME, '../src/db/migrations');
@@ -29,6 +31,7 @@ const client = new Kysely<DB>({
       connectionString: container.getConnectionUri(),
     }),
   }),
+  log: kyselyLogger,
 });
 
 vi.mock('../src/db/database.ts', () => {
